options-trading-web/backend/requirements.txt
# FastAPI框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.0
redis==5.0.0

# 任务队列
celery==5.3.0

# 数据获取
httpx==0.26.0
aiohttp==3.9.0

# 工具
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.0
python-multipart==0.0.6

# 科学计算
numpy==1.26.0
scipy==1.12.0
pandas==2.2.0

# 测试
pytest==8.0.0
pytest-asyncio==0.23.0
